feat: extract article category component and mock article categories data
