--- src/components/pages/ArticleCategories.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-type Props = {}
-
-export default function ArticleCategories({}: Props) {
-  return <div className="w-52 bg-white">ArticleCategories</div>
-}
