+++ src/components/pages/ArticleCategorySection.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import AriticleCagegory from '../shared/AriticleCagegory'
+
+import articleCategories from '../../mockData/articleCategories'
+
+type Props = {}
+
+export default function ArticleCategorySection({}: Props) {
+  return (
+    <div className="w-52 p-4 bg-white rounded-lg">
+      <AriticleCagegory categories={articleCategories} />
+    </div>
+  )
+}
