+++ src/components/pages/ArticleSection.tsx
@@ -3,5 +3,7 @@ import React from 'react'
 type Props = {}
 
 export default function Articles({}: Props) {
-  return <div className="mx-12 bg-white grow max-w-180">Articles</div>
+  return (
+    <div className="mx-12 bg-white grow max-w-180 rounded-lg">Articles</div>
+  )
 }
