+++ src/components/pages/HomeContent.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import ArticleCategories from './ArticleCategories'
+import ArticleCategorySection from './ArticleCategorySection'
 import ArticleSection from './ArticleSection'
 
 type Props = {}
@@ -7,7 +7,7 @@ type Props = {}
 export default function HomeContent({}: Props) {
   return (
     <div className="my-8 mx-36 flex">
-      <ArticleCategories />
+      <ArticleCategorySection />
       <ArticleSection />
     </div>
   )
