+++ src/components/pages/SidebarMobileContainer.tsx
@@ -1,35 +1,9 @@
 import React, { useState } from 'react'
 import { MenuIcon } from '../../icons'
-import animal from '../../images/animal.jpg'
-import chat from '../../images/chat.jpg'
-import food from '../../images/food.jpg'
-import popular from '../../images/popular.jpg'
-import travel from '../../images/travel.jpg'
 import Button from '../shared/Button'
 import SidebarMobile from './SidebarMobile'
 
-const mockCategories = [
-  {
-    title: 'popular',
-    imgSrc: popular,
-  },
-  {
-    title: 'chat',
-    imgSrc: chat,
-  },
-  {
-    title: 'animal',
-    imgSrc: animal,
-  },
-  {
-    title: 'food',
-    imgSrc: food,
-  },
-  {
-    title: 'travel',
-    imgSrc: travel,
-  },
-]
+import articleCategories from '../../mockData/articleCategories'
 
 type Props = {
   isLogin: boolean
@@ -54,7 +28,7 @@ export default function SidebarMobileContainer({ isLogin }: Props) {
   return (
     <SidebarMobile
       isLogin={isLogin}
-      categories={mockCategories}
+      categories={articleCategories}
       isOpened={isOpened}
       onDismiss={handleToggle}
       Toggle={Toggle}
