+++ src/mockData/articleCategories.js
@@ -1,0 +1,30 @@
+import animal from '../images/animal.jpg'
+import chat from '../images/chat.jpg'
+import food from '../images/food.jpg'
+import popular from '../images/popular.jpg'
+import travel from '../images/travel.jpg'
+
+const mockCategories = [
+  {
+    title: 'popular',
+    imgSrc: popular,
+  },
+  {
+    title: 'chat',
+    imgSrc: chat,
+  },
+  {
+    title: 'animal',
+    imgSrc: animal,
+  },
+  {
+    title: 'food',
+    imgSrc: food,
+  },
+  {
+    title: 'travel',
+    imgSrc: travel,
+  },
+]
+
+export default mockCategories
